refactor: rename folder plugins -> plugin
