+++ include/open62541pp/open62541pp.hpp
@@ -11,11 +11,11 @@
 #include "open62541pp/monitoreditem.hpp"
 #include "open62541pp/node.hpp"
 #include "open62541pp/nodeids.hpp"
-#include "open62541pp/plugins/accesscontrol.hpp"
-#include "open62541pp/plugins/create_certificate.hpp"
-#include "open62541pp/plugins/log.hpp"
-#include "open62541pp/plugins/nodestore.hpp"
-#include "open62541pp/plugins/pluginadapter.hpp"
+#include "open62541pp/plugin/accesscontrol.hpp"
+#include "open62541pp/plugin/create_certificate.hpp"
+#include "open62541pp/plugin/log.hpp"
+#include "open62541pp/plugin/nodestore.hpp"
+#include "open62541pp/plugin/pluginadapter.hpp"
 #include "open62541pp/result.hpp"
 #include "open62541pp/server.hpp"
 #include "open62541pp/services/services.hpp"
