+++ include/open62541pp/plugin/accesscontrol.hpp
@@ -6,7 +6,7 @@
 #include "open62541pp/bitmask.hpp"
 #include "open62541pp/common.hpp"  // AccessLevel, WriteMask
 #include "open62541pp/detail/open62541/common.h"  // UA_AccessControl
-#include "open62541pp/plugins/pluginadapter.hpp"
+#include "open62541pp/plugin/pluginadapter.hpp"
 #include "open62541pp/span.hpp"
 #include "open62541pp/types.hpp"
 #include "open62541pp/types_composed.hpp"  // UserTokenPolicy, PerformUpdateType
