+++ include/open62541pp/plugin/pluginadapter.hpp
