+++ include/open62541pp/server.hpp
@@ -10,7 +10,7 @@
 #include "open62541pp/config.hpp"
 #include "open62541pp/detail/open62541/server.h"
 #include "open62541pp/nodeids.hpp"
-#include "open62541pp/plugins/log.hpp"
+#include "open62541pp/plugin/log.hpp"
 #include "open62541pp/span.hpp"
 #include "open62541pp/subscription.hpp"
