+++ src/client_config.hpp
@@ -7,8 +7,8 @@
 #include "open62541pp/wrapper.hpp"
 
 #include "customdatatypes.hpp"
-#include "plugins/log_default.hpp"
-#include "plugins/pluginmanager.hpp"
+#include "plugin/log_default.hpp"
+#include "plugin/pluginmanager.hpp"
 
 namespace opcua {
