+++ src/plugin/create_certificate.cpp
@@ -1,4 +1,4 @@
-#include "open62541pp/plugins/create_certificate.hpp"
+#include "open62541pp/plugin/create_certificate.hpp"
 
 #ifdef UA_ENABLE_ENCRYPTION
 
@@ -9,7 +9,7 @@
 #include "open62541pp/config.hpp"
 #include "open62541pp/detail/open62541/common.h"
 #include "open62541pp/exception.hpp"
-#include "open62541pp/plugins/log.hpp"
+#include "open62541pp/plugin/log.hpp"
 #include "open62541pp/typewrapper.hpp"
 
 #include "log_default.hpp"
