+++ src/server.cpp
@@ -12,8 +12,8 @@
 #include "open62541pp/event.hpp"
 #include "open62541pp/exception.hpp"
 #include "open62541pp/node.hpp"
-#include "open62541pp/plugins/accesscontrol.hpp"
-#include "open62541pp/plugins/nodestore.hpp"
+#include "open62541pp/plugin/accesscontrol.hpp"
+#include "open62541pp/plugin/nodestore.hpp"
 #include "open62541pp/services/attribute_highlevel.hpp"
 #include "open62541pp/session.hpp"
 #include "open62541pp/types.hpp"
